+++ .github/workflows/CI_RAFT.yml
@@ -16,7 +16,7 @@ jobs:
             fail-fast: false #true
             matrix:
                 os: ["ubuntu-latest", "macOS-latest", "windows-latest"]
-                python-version: ["3.9", "3.10", "3.11"]
+                python-version: ["3.10", "3.11"]
 
         steps:
             - name: checkout repository
